cryptoquant-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
httpx>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.9
google-genai>=0.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
deep-translator>=1.11.4
